Add a rule to clarify vertical stagger suffixes (#65)
This PR solves issue #61 on clarifying the use of the _at_interface suffix. I expanded rule number 4, which describes the use of location suffixes:

By default (when not specified otherwise), variables are grid means or centers (defined by the host). If a variable is defined at a different physical location, a qualifier should be used to denote this. For example, to specify the vertical location of a variable with respect to vertical grid cells, the following rules apply. If [variable] doesn't have a vertical location suffix and is defined on n levels:

[variable]_at_interfaces is defined at the interfaces between grid cells vertically, including the bottom-most and top-most interfaces. It is defined on n+1 levels.
[variable]_at_top_interfaces is defined at the interfaces between grid cells vertically, including the top-most interface but excluding the bottom-most interface. It is defined on n levels.
[variable]_at_bottom_interfaces is defined at the interfaces between grid cells vertically, including the bottom-most interface but excluding the top-most interface. It is defined on n levels.