Update StandardNamesRules.rst: change`surface` prefix  to `_at_surface` suffix (#72)
This PR changes the rules (and the rules only) to make surface a suffix (at_surface) rather than a prefix. This proposed change was discussed in last week's CCPP framework meeting and is a departure from the CF conventions, which use surface_ as a prefix. The motivation for this change is to maintain consistency with all other level qualifiers that are used as _at_level qualifier (i.e. as suffix).